refactored index.js and offer-details
